Jupyter Notebook
In this project you will have to put yourself in the shoes of a data scientist and identify the factors contributing to attrition of the employees and recommend possible solutions. The files contain complete staff utilization reports for all employees of the XYZ Corp. through 2016-2018. The data contains details of employees as well as the details of the projects that they have worked on in the last two years. The other file contains all the attrition in the organization for the years 2015-18 with details such as reason of attrition along with other employee details. You are free to use external data sets which may impact the attrition.